fix: Make default classification_level "Internal" str (#14)
Bugreport from the "convert-ascii" command stated that the .nc file
outputted did not automatically pass validation due to
"classification_level" not being set to the enum string member value. It
instead took the representation of the enum member, as seen here:
![Screenshot 2023-11-24 at 15 58
56](https://github.com/equinor/atmos-validation/assets/44391464/915c5c7e-9331-4b25-b18f-f40123edc118)

Apperantly using the enum as the default value triggers this behaviour,
so just switched it "Internal" instead.