feat: update parameters config with qc_test (#40)
This PR adds a new schema for a QCTest, that describes a test to be used
during QC process for a given parameters.

Each Parameter can have multiple QCTests associated with it. A QCTest
can also be used on different parameters (though with potentially
different default values in the default_parameters list). The QCTest
implementations are provided by metocean. The default parameter values
are provided by persons with admin privileges.