feat: update qc test schema (#42)
This PR renames `test_name` to `long_name `since this is what it’s
called in the `metocean` package, and is in-line with `long_name` field
for `Parameter`s.

It also renames `default_parameters `to `default_variable`s to avoid
confusions with the atmos `Parameter`s

It also adds new types to the `default_variables` to accommodate the
different types required by the q -tests in the `metocean` package

It also changes the` List[QCTest]` to `Dict[str, QCTest]` in the
`ParameterConfig schem`a, so that the tests associated with a
`Parameter` can be searched by id. This will be useful in future apis
which need to modify and delete qc-tests from a `parameter` instance.